+++ example/android/app/src/main/kotlin/com/example/pinch_zoom_any_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.pinch_zoom_any_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()